optimize indexing dataloader

+++ scvi/models/utils.py
@@ -4,7 +4,7 @@ from torch.distributions import register_kl, Multinomial
 
 def one_hot(index, n_cat):
     onehot = torch.zeros(index.size(0), n_cat, device=index.device)
-    onehot.scatter_(1, index, 1)
+    onehot.scatter_(1, index.type(torch.long), 1)
     return onehot.type(torch.float32)
